Refactor: userStore.js 위치 변경
